refactor s3utils package
