+++ s3/helpers_test.go
@@ -19,6 +19,7 @@ import s3mocks "github.com/RedHatInsights/insights-operator-utils/s3/mocks"
 var (
 	testBucket             = "my_bucket"
 	testFile               = "my_file"
+	testFileCopy           = testFile + "_copy"
 	fileContent            = "some content"
 	randomError            = "an error"
 	testFileInvalidContent = "invalid_content"
@@ -36,6 +37,7 @@ type testCase struct {
 	errorMsg        string
 	mockContents    s3mocks.MockContents
 	file            string
+	files           []string
 	body            []byte
 	downloadError   error
 	wantFiles       []string
